\name{aplpack-package}
\alias{aplpack-package}
\alias{aplpack}
\docType{package}
\title{
Another PLotting Package 
}
\description{
 this package contains functions for plotting: chernoff faces, stem and leaf plots, bagplot etc.
}
\details{
\tabular{ll}{
Package: \tab aplpack\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2006-04-04\cr
License: \tab GPL \cr
}
   bagplot plots a bagplot
   faces plots chernoff faces
   scboxplot plot a boxplot into a scatterplot
   spin3R enables the inspection of a 3-dim point cloud
   stem.leaf plots a stem and leaf display
}
\author{
Peter Wolf
Maintainer: Peter Wolf <pwolf@wiwi.uni-bielefeld.de>
}
\references{
}
\keyword{ package }
\examples{
  
}
